# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "BKRES API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/gateway/create": {
      "post": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to create gateway",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "gateway_name": {
                  "example": "any"
                },
                "lat": {
                  "example": "any"
                },
                "lon": {
                  "example": "any"
                },
                "is_public": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/gateway/deletegateway/{API}": {
      "delete": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to delete gateway",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/gateway/updateGateway/{API}": {
      "put": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to update gateway",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "gateway_name": {
                  "example": "any"
                },
                "lat": {
                  "example": "any"
                },
                "lon": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/gateway/getGateway/{API}": {
      "post": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to get gateway",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/gateway/getallGateway": {
      "post": {
        "tags": [
          "Gateway"
        ],
        "description": "Endpoint to get all gateway",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sortBy": {
                  "example": "any"
                },
                "sortOrder": {
                  "example": "any"
                },
                "filterKey": {
                  "example": "any"
                },
                "filterValue": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/sensor/create": {
      "post": {
        "tags": [
          "Sensor"
        ],
        "description": "Endpoint to create sensor",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "device_id": {
                  "example": "any"
                },
                "sensor_name": {
                  "example": "any"
                },
                "provider": {
                  "example": "any"
                },
                "unit": {
                  "example": "any"
                },
                "describe": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/sensor/deletesensor/{API}": {
      "delete": {
        "tags": [
          "Sensor"
        ],
        "description": "Endpoint to delete sensor",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/sensor/updateSensor/{API}": {
      "put": {
        "tags": [
          "Sensor"
        ],
        "description": "Endpoint to update sensor",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sensor_name": {
                  "example": "any"
                },
                "device_id": {
                  "example": "any"
                },
                "provider": {
                  "example": "any"
                },
                "unit": {
                  "example": "any"
                },
                "describe": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/sensor/getSensor/{API}": {
      "post": {
        "tags": [
          "Sensor"
        ],
        "description": "Endpoint to get sensor",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/sensor/getallSensor/{API}": {
      "post": {
        "tags": [
          "Sensor"
        ],
        "description": "Endpoint to get all sensor",
        "parameters": [
          {
            "name": "API",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  }
}